View Variables Option:  Add/Remove Verb
Since we can't edit the verbs variable directly this allows us to do so safely.

It builds a list of procs and verbs for the mob type and allows the admin to add it.

Want to add self_destruct to a cyborg?  It will show up in Commands tab
Want to add RevConvert to a regular Rev?  Feel free!

Want to remove whisper from some loud mouth?  Go wild!

Adding procs that aren't usually available to players will create a commands tab, primarily this is for debugging since Advance ProcCall on mobs is awful.

I figured this would be useful when I noticed traitorpanel added RevHeads didn't get their RevConvert verbs and I could find no way of adding it otherwise.