DNA2 rangebug, if any hair/beard/eye RGB values are 0 it won't encode it into the dna.
This is a hacky fix and is temporary until I can sit down and figure out exactly why it's not encoding that block of DNA properly.

The result of it not encoding it into the dna is that if you clone the person they have a random value there that was assigned at round start.