Mentor Unnerf.
JMP was doing a rights check in the Topic instead of doing it on the proc, just went ahead and fixed the rights check to include mentors.
Mentors now have player panel but can't see traitor panel.  The idea is so they can't tell who is an antag through this.