Teleporter mapping enhancement
Allows teleporters to be mapped in any direction. Set the dir of the teleporter computer to the direction that the hub is in. I.e. if hub is east, set tele comp dir to 4 (east).
Teleporter computer sets up refs to the hub and station so we don't have to search for the other teleporter components every time we teleport.
Applies underlays and overlays to the components to correctly align the connecting cables.