fixes vox armalis being a choosable race
the IS_WHITELISTED flag will add the species to the whitelisted_species list,
which in turn gets used as the base list of choosable species in the char setup when
the server config has species whitelisting off.

unless you want giant combat spessbirds as crew ( as a security force maybe ? ;) ),
this is a bug.