Made spawning with a loyalty implant compliant to configuration...
Created a proc for being implanted with loyalty implant which
checks for the configuration for whether or not it will work. If it is
enabled, it continues on-- if it is disabled, it returns from the proc.

Also made secure closets not automatically spawn them unless the config
is set to allow it.

Admin 'loyalty implanting' via the panel is now compliant to the
configuration file. If preferred, I am willing to make an override variable
to the proc that will allow admins to do it anyway.

Supply pack for it commented out due to inability to make it conform.
Easy enough to re-add back for anyone downstream if they wish it. Same
applies to the design.


Signed-off-by: Decius <deciusreln97@gmail.com>