fixes #6023 and cleans up mechbay.dm
- adds tile update on attackby() to fix #6023 using a so far dormant icon state ( line 49 )
- adds missing console initalization on attack_hand() ( line 189 )
- comments unreachable ( via normal means ) mech_bay_power_console/Topic ( line 225 )
- gets rid of relative paths, leaving procs unchanged
note : recharge-floors currently don't care about what direction the console is, unlike the charger-port which always has to be west of the floor.
this should be probably changed to console having to be east of the floor, as most mappers do that anyway and to reduce the chance of future bugs. awaiting feedback on that one.
 like the charger and floor.