Updates storage costs
Storage cost of an item is now twice that of the next smallest item, instead of being equal to the w_class.
Items with w_class 1 or 2 are unaffected. This produces much better behaviour for containers that are limited
by storage space instead of slots, such as how toolboxes are planned to work in a future update.