Adds Icon Caching for some Mob HUD Icons
Adds partial caching for some of the smaller HUD additions of the mobs to allow for less processing needed to constantly update the icons when needed.
Code has been tested on a live server for several weeks prior to it's shutdown, so it should work as great as it has worked for us.