Fixes #10309
Antag types with ANTAG_OVERRIDE_JOB are finalized before jobs are handed out, based on the idea that something which replaces a player's job should not need to depend on their assigned job in any way.
Moves the selection of pending antags outside of can_start().
Fixes issue with antag distribution that occurs when a player is a candidate for multiple antagonist types.