Fixes observers being unable to see silicon antag status.
target_list contains mob/living mobs while the previous loop only looped over living/carbon/human.
To make things a bit of a headache hud_list is defined separately for living/carbon/human and living/silicon, instead of directly on /living.

It is certainly possible to define hud_list at mob/living instead but adding an extra loop for silicons is the method that is least likely to break things, thus the most appropriate for a master fix.