Revert to old_x and old_y after attack animations (#85)
Does not use initial(pixel_x) and initial(pixel_y) for 'finishing' attack animations, so that mobs can have old_x and old_y updated to reflect a natural pixel offset (e.g. for mobs that are longer/taller than 32x32 and are nudged to fit into a tile).

For example, we have a 32x64 cyborg module which sets pixel_x, pixel_y, old_x, and old_y when selected, however attacking with it resets it to 0 as this doesn't respect old_x and old_y like all other mob anims do (see floating, etc).

Also why define this on atom and then literally never use it on anything other than a mob, in the... mob/animations.dm file?

_Sorry to keep making QOL PRs. I promise I'll send something cool eventually._