Fixes #1655
Don't snowflake one mode with new(). Setting this on new() causes
weirdness when new() is called (prints unexpected messages to players
about the mode being toggled). If you really want it, define the var as
equal to the bitflag in the definition and update the starting human HUD
to match. Don't change it 'after' if that's always how you want it.