Fixes #1714
Removes duplicate code, and fixes the logic of the original code that was supposed to perform the same function, but did not, due to bad if-statement logic. Performed eye removal, and eye transplant to test. I would have liked to have fixed the flat `blinded = 0` at the top of the file, but it's so ingrained in how it works now, I could not.