Adds in taur specific void/rigsuits (#524)
* Code works, North & South works.

East and West still requires fixing, however.

* Has west sprites working, east still broke

* Another fix

* Reverts taursuits sprite adjustment, changing

it to just adjust the X variable

* Makes on floor sprite look correct

Updates icon when worn, too.

Still need to figure out how to set on-sprite pixel_16 to -16, though.

* Typo fix

* This works

The /taur voidsuits work like a charm.
With the rigsuits, it'll show the correct sprite for taurs while the
rigsuit is sealing, but after it seals up it'll revert to the old
sprite. To fix this, all someone has to do is toggle and retoggle the
chestpiece.

If someone else without a taur/a different taur goes to wear a
rigsuit/voidsuit someone previously wore it'll change to fit their taur.
Pretty neat!

Sprites provided by SilencedMP515
Also, thanks to Nightwing for helping me get this working when I ran
into a brick wall.

* Fixes