Adds KH life crystal fluff/loadout item
It just exists as a fluff excuse as to how you get cloned offsite if you don't get cloned at CC, if you want an alternative. The little lore behind it is that it notifies an offsite agency if you die so you can get cloned by them. It has a little paper in the carrying case explaining that.

It's a uniform accessory, or you can just pocket it. They are unmarked, so finding a broken one shouldn't be very interesting evidence or influence much outside just the fluff idea of their existence.

(Also, it's a crystal, but no it's not magic. The manual goes into detail about the power source, transmitter, etc.)