Fixes a problem with advanced_who
Advanced who was doing (Testname represents Ckey, StealthTest represents stealth Ckey)
"Testname" for admins when unstealthed, which was good, but it was doing
"TestnameStealthTest" when admins were stealthed.

This PR makes it start off as "Testname", then checks to see if they are stealthed. If so, it becomes "StealthTest," then adds in all the additional info (If they are observing, in lobby, or playing.)