Fed and KHI Uniforms/Suits
Adds Joan's Fed overcoats and my KHI uniforms as optional alternate uniforms for departments. Each requires you to be in that department to let you spawn with it.

(The Kitsuhana.dmi change is a change to my own synth sprite to fix a sprite error, unrelated)
(Also updates the sprite in custom_clothes_vr.dmi for an item for Tempest for her medhud)