Whitelist and Custom Items gitignore update
I *really* don't care if players can see these on the server Github. Not
only will this make things easier to update, it'll also stop this stupid
bullshit idea that the server is full of snowflakes with gun permits.
Full transparency!

Yes I'm aware we could just use filezilla to do this but screw that.