Nerfed Xenoarch Space Suit Slowdown
Seriously, why the hell is this suit such a nightmare to move around in?
Mechanically, it has ZERO armor besides bio/radiation protection, and
the only real perk of it is that it's space worthy.

Xenoarcheology is just tedious and annoying with these dumb suits being
so clunky, so I've eliminated the slowdown to only slow you down as much
as voidsuits.