Adds defibrilator kit
Adds a defib kit. Not ported from /tg/, not a backpack or anything. It's more like an AED than anything. Click with empty hand or drag onto yourself to take the pads out, and put them into some dead person's uniform slot.

Currently has 3 attempts with 75% chance of success with the default battery.

The paddles are cabled, so if you walk away with them it yanks the cable out and you have to reconnect them. Patient can't be moving either, and it won't shock living people (someone can code an emag thing for that I guess?).

(Excludes .dme addition)