Improves bandanas
No more tie verb. Just put it into whatever slot and it (correctly) adjusts the needed stuff to work. Added 'Can be worn on the head or face' to the description to indicate that.

Currently wearing a bandana on your head makes it impossible to eat, which doesn't make sense, so this fixes that too.