Adds in prey transfer (and a bug)
Adds in prey transfer.

Has a bug in which if the pred deletes their tranfer location stomach,
then the prey will be able to be moved there, trapping them in limbo.

Shouldn't happen often if ever, but I still need to figure out how to
fix it.

Merging this in case anyone has a solution of how to do the sanity
check.

Lines 416-428 in belly_vr.dm