Increase transfer rate of Xenobio injector
It was not coded correctly. This "transfer_amount" variable is never set anywhere. The "trans_to_mob" reagent transfer function defaults to 1 if you don't give it something, and this proc passes the null transfer_amount to it, which is basically that. So it always transfers 1. 5 is easier to live with.