Toilet fixes
Yes yes. I know.

Anyway, there's no cooldown on this so you can give someone about 800 brute in 10 seconds. A LITTLE UNUSUAL if you ask me. Also reduced it from 8 because what the hell is that toilet seat made out of? Osmium? For an armor-ignoring attack that's a little high. Also the swirlie thing was coded wrong for the do_after.