Give internal cell it's own icon
Was a little weird to reuse the supercell... there are already WAY TOO MANY "not an actual powercell" powercells. So this one is just repainted blue and joins all the other prosthetic organs in surgery.dmi. The icon line can just be removed because the parent points to this file.