Handheld Translator
Turn it on, keep it in hand or pocket, and it'll translate languages for you (not signlanguage...).
It doesn't bother displaying languages that you speak. I assume your character doesn't look down at the screen if they understood what was said (it's mostly just to prevent spam when only 1 in 4 people standing around talking is speaking another language).

There's one on a desk in the security locker room and one in the medical breakroom. I figure these two departments have the highest need of translations. R&D can make more. The design is fairly cheap.