Allow MMI/Posi/Digi brains w/o minds to be put in
Right now if you try to put an empty posibrain into a mob it won't let you. But with resleeving, this might be just the step before putting a mind in them, so that's sort of not how it should work. When it was just a ghost-trap object that ghosts could get into, maybe, but it's moved beyond that.