Angel and Demon languages
Not actually MAGICAL angels and demons, I believe, just they resemble them and are spacefaring races and so sometimes end up matching various religious texts in terms of appearance and habit.

Can we please stop adding whitelisted languages? It's dumb. We have ONE person whitelisted for gutter, and a lot of good it does them since nobody else has even applied. They'd have to talk to someone with a translator.