Adds the 'Hunter' Capture Gun
Dual mode stun-beam and net launcher, craftable in R&D. Would be legal for security to have on green since it's non-lethal. The stun beam has 10 shots like the taser, and the net launcher has 2 (uses the same energy). So basically 5 stun+1 net, or 2 nets, or 10 stuns.

Uses a new overlay system for updating it's icon rather than the "have one state for every possible combination" to prevent a huge mess of iconstates. Maybe Joan could look at this and use it sometime.