Adds more transformation types for specific
Changes "Species Transformation" to "Replica Transformation" for
clarity, as people were using it and expecting them to change species,
not be a perfect 1:1 replica of the person doing the TF.

Adds "Change_species_and_taur" mode, which does exactly as it says
without making the person appear as a 1:1 replica as the pred TFing them
Adds "Change_hair_and_eyes" tf mode, which allows the changing of the
prey's hair and eye color without having to completely TF them or change
their gender.

Fixes a bug that infinitely drained the pred's nutrition when they used
a TF mode. Slightly increases amount of nutrition the prey gets from the
pred.