NSFW Gun
Also fixes slidey walls at centcom, gives the ERT commander a NSFW with stun, xray, lethal, and net. Gives deathsquad a full NSFW kit (including 'final option' ammo), replaces my own fluff Hunter with a NSFW with stun and net (so, same as the hunter, but that sprite is so badass I can't help it), and fixes a typo on my permit while I'm on there (KIN isn't a thing, KHI is).