Fixes for nettles (#3854)
* Fixes for nettles
Nettle plants can now spawn randomly from weeds - it was trying to spawn 'nettles', but the name of the plant is 'nettle', so it never spawned. Now it can.

You now hit people with food if you use it on harm intent. This is mostly so that you can use nettles as a weapon again, but hey, who wouldn't want to hit people over the head with a baguette.

Nettles and other stinging "fruits" can no longer be used to generate infinite chemicals by throwing them. Chemicals that the "fruits" inject are now properly removed when hitting a target.

* Fixes for nettles
Nettle plants can now spawn randomly from weeds - it was trying to spawn 'nettles', but the name of the plant is 'nettle', so it never spawned. Now it can.

You now hit people with food if you use it on harm intent. This is mostly so that you can use nettles as a weapon again, but hey, who wouldn't want to hit people over the head with a baguette.

Nettles and other stinging "fruits" can no longer be used to generate infinite chemicals by throwing them. Chemicals that the "fruits" inject are now properly removed when hitting a target.