Drone circuit +posibrain lorechanges
Removes all that mumbo about "you live to serve" and replaces it with a reminder that you're designed to be friendly to human values. I don't really see the trouble with this-- if a new drone or posi wants to be a surly bastard, it can, it's just not getting sleeved in a shiny bipedal body.

Also adds a bunch of indulgent OS names for drone circuits.

(Realtalk this is mostly just me figuring out how PRs work so I can start being something closer to actually productive)