Sanitize sprite sheet species call
Fixes #2186 , this bug was not reported in Polaris, and its kinda critical due to its grieffy nature, so its best to first patch it, and then get someone to comment it upstream.

Special thanks to Bluehaus for testing the fix, and to Mewchild for reporting and testing the bug in the first time.